converted some byte to short

+++ code/code/cmd/cmd_set.cc
@@ -505,7 +505,7 @@ void TPerson::doSet(const char *argument)
         amt *= parm2; 
         amt /= 10000;
         if (discArray[snt]->startLearnDo > amt) {
-          amt = min(discArray[snt]->startLearnDo, mob->getMaxSkillValue(snt));
+          amt = min((short)discArray[snt]->startLearnDo, mob->getMaxSkillValue(snt));
         }
         mob->setSkillValue(snt, max(amt, 1));
         mob->setNatSkillValue(snt, amt);
